Allow specs to be represented by a simple string
In most cases, a URL should be all we need in that list. Getting rid of the
wrapping `{ "url": ... }` struct in these cases allows to keep the list more
human readable.

This update allows entries to be just a string. Linting code updated to prefer
the string format when there is only a URL (which is going to be the case as
long as we don't create additional properties). Specifications remain sorted
by alphabetic order of their URL.

The `index.js` file continues to export an array of objects with a `url`
property to ease ingestion by external tools.

See discussion in:
https://github.com/w3c/browser-specs/pull/2#pullrequestreview-345692753

(Note that if we actually discover later on that a URL is **not** enough in
most cases, we may want to undo this change...)