Add logic to compute shortname, family name and level
Most specification URLs follow a few common patterns from which it is easy to
derive a shortname (e.g. "webauthn-2" from "https://www.w3.org/TR/webauthn-2/").
In turn, this shortname can be used to derive a level ("2" for "webauthn-2") and
a family name that is common across levels of the spec ("webauthn" here).

The "compute-shortname.js" module does just that. It handles main types of URLs
and returns an object with "shortname", "familyname", and a "level" property
when that makes sense.

An exception is thrown when the shortname cannot be computed. When that happens,
the corresponding entry in "specs.json" should be updated to make the shortname
(and/or family name and/or level) explicit. For instance, the entry for CSS
Media Queries Level 3 should typically read:

```json
{
  "url": "https://www.w3.org/TR/css3-mediaqueries/",
  "familyname": "mediaqueries",
  "level": 3
}
```

The list of specs exposed in index.js now features the "shortname",
"familyname" and "level" properties for each spec.

The family name is typically the property that Reffy would use to name CSS/IDL
extracts. Testing on the list of specs used by Reffy (not part of this update),
this seems to work just fine.

The update comes with unit tests.