Add logic to compute shortname, family name and level (#7)
Most specification URLs follow a few common patterns from which it is easy to
derive a name (e.g. "webauthn-2" from "https://www.w3.org/TR/webauthn-2/").
In turn, this name can be used to derive a level ("2" for "webauthn-2") and
a shortname that is common across levels of the spec ("webauthn" here).

The "compute-shortname.js" module does just that. It handles main types of URLs
and returns an object with "name", "shortname", and a "level" property when that
makes sense.

An exception is thrown when the name cannot be computed. When that happens,
the corresponding entry in "specs.json" should be updated to make the name
(and/or shortname and/or level) explicit. For instance, the entry for CSS Media
Queries Level 3 should typically read:

```json
{
  "url": "https://www.w3.org/TR/css3-mediaqueries/",
  "shortname": "mediaqueries",
  "level": 3
}
```

The list of specs exposed in index.js now features the "name", "shortname" and
"level" properties for each spec.

The shortname is typically the property that Reffy would use to name CSS/IDL
extracts. Testing on the list of specs used by Reffy (not part of this update),
this seems to work just fine.

The update comes with unit tests.