Allow delta flag and add previous/next links
To make sense of a given spec, it is necessary to track whether it is a "delta"
spec that may not contain the entire contents of the previous level or a "full"
spec. This update allows specs to be defined as delta specs.

Whether the spec is a delta spec cannot be determined automatically. Delta specs
need to be flagged with a "delta" flag in specs.json through a "delta" property
as in:

  {
    "url": "https://www.w3.org/TR/mydelta-2/",
    "delta": true
  }

To keep the list of URLs in specs.json readable, the flag can also be specified
directly within the string that describes the spec, after the URL:

  "https://www.w3.org/TR/mydelta-2/ delta"

This notation is syntactic sugar for the previous notation and can only be used
when a string is used to describe the specification. The linter will actually
prefer the synctactic sugar notation when possible (meaning when no other
property is defined).

To ease traversing operations between levels of the same specifications, the
list of specs returned by "index.js" now features a "previousLevel" and
"nextLevel" properties, when possible, set to the name of the spec that
represents the previous and/or next level. By construction, all specs linked
together in a chain of previous/next links share the same "shortname" property.

The "previousLevel" property is typically useful to identify the name of the
previous level that contains the base contents of a delta spec.

From a code perspective:
- The linter was updated to handle the delta flag, notably the syntactic sugar
notation, and to make sure that delta spec definitions are ok (delta specs
should always have a previous non delta level in the list)
- A new compute-prevnext module was created to compute previous/next links
- The index.js code was updated to include previous/next links in the list of
specs returned
- The JSON schema for specs.json was updated to add the "delta" flag and renamed
to "specs-schema.json". The "multipleOf" condition was also dropped (because
0.1 multiples cannot work because JS only has floats).
- A new JSON schema was defined to check the list of specs returned by
"index.js". It is similar to "specs-schema.json" but also has "previousLevel"
and "nextLevel" properties.
- New tests were added for the linter, the new module, and for the list of
specs returned by index.js.