Add currentLevel property (#18)
See #9.

Groups may work on multiple levels of a given spec at once, and may want to
redirect readers to one of these levels by default.

With this update, the code exposes a new "currentLevel" property for each spec
in the list that points at the name of the specification with the same shortname
that should be viewed as the current one.

By default, the "currentLevel" property points at the latest level that is not
a delta spec. This rule may be overwritten by flagging the entry that should
be seen as the current level in specs.json. This can be done through a
"forceCurrent" boolean property:

```json
{
  "url": "https://www.w3.org/TR/spec-1",
  "forceCurrent": true
}
```

... or inline next to the URL with a "current" keyword:

```json
"https://www.w3.org/TR/spec-1 current"
```

The linter automatically switches to the latter form when possible, and
complains when:
1. a delta spec is flagged as the current level
2. the list contains more than one spec flagged as current for a shortname

The linter also automatically drops current flags that are useless, e.g. because
the flag is added to the latest spec that is not a delta spec.