Export completed list of specs in index.json file
See discussion in #13.

This update replaces the `fetch-info` workflow with a `build` workflow that
directly creates the index file `index.json` that contains the exposed list of
specs generated from the `specs.json` file.

This update also drops tests that compare the `index.json` file with the
`specs.json` file, to allow for pull requests that only change `specs.json`
(Impact on the generated list would still need to be assessed somehow, see #17).

Note the code preserves the index.js file which is still a useful CLI.