Add code to manage releases
Idea is to use release-it to manage releases. With the update, a new release
should be doable through the following command-line:

  npm run release

Command must be run from the `master` branch. A `GITHUB_TOKEN` env variable
needs to be set to a GitHub Personal Access Token with `repo` rights.

The release-it library will take care of enforcing that the local folder looks
good (no pending changes) and that tests pass before releasing the new version.
It will create a new commit "vX.Y.Z", a new tag, a new GitHub Release with a
changelog, and publish the package to npm.

The command is interactive and asks questions such as "hey, what should the new
version number be?" with choices (major update, minor update, patch, etc.).

Of course, the command is only going to succeed provided that the user who runs
it has the appropriate admin rights on the GitHub repo and on the npm package.