Rename xxxInSeries to seriesXxx (#34)
This update renames `previousInSeries` and `nextInSeries` to `seriesPrevious`
and `seriesNext` to improve naming consistency with other series properties,
which all start with `series`.

Note: it is tempting to put these properties under the `series` object, but that
object is meant to convey information about the series itself, not about the
current spec in the series, which `seriesComposition`, `seriesVersion`,
`seriesPrevious` and `seriesNext` hold. In other words, the `series` object is
the same for all specs in the series.