Add changelog and mechanism to update it on release (#37)
The CHANGELOG.md file will be included in the NPM package by default. It will
link to commits and pull requests.

Note: the hardcoded URLs in .release-it for `--commit-url` et al. are needed not
to force an origin name on everyone (personally, I use `w3c`, but I suspect
others may either stick to `origin` or use something else).