Linter: separate tests from actual linting code
The linter was doing two things at once:
1. running validity tests against specs.json
2. linting specs.json (or reporting on possible need to lint it)

In practice, this meant that running `npm test` would not test `specs.json`,
which does not seem to be a good thing.

The update puts validity tests in a proper test module. The update will also
allow to run the linter as an automated job on each and every push.