Adjust a couple of TR URLs to match W3C API shortnames (#59)
Recent release of the W3C API includes adjustments to default shortnames for
some of the specs. A couple of updates were needed for the CSS Images Module
Level 3 and Trace Context.

The W3C API returns a 301 redirect when a non default shortname is used, which
the code in `fetch-info.js` currently does not support. That should probably be
fixed, but it seems a good thing in any case to align default shortnames with
those of the W3C API.