Add tests to compare base URL with release/nightly URL
The base URL should either match the release URL if there is one or the nightly
URL otherwise. New tests will fail when there are entries in `index.json` for
which that is not the case.

The first test should also help with identification of situations where a FPWD
was published, and is known to Specref but not to us (#62).

Note that, as opposed to other tests that only depend on internal code logic,
the second test may fail because external sources (Specref in practice) need to
be fixed to return the right nightly URL. Fixing the problem at the source may
take some time.