Expose list of pages for multipage specs (#123)
This udpate allows to flag multipage specs in `specs.json` (inline or through
a `multipage` boolean). This flag instructs the build code to go and fetch the
index page and extract the list of pages.

The list of pages gets exposed in `release.pages` and `nightly.pages`.

Note that flagging the specs merely allows to restrict the number of specs that
get fetched. In theory, the code that extracts pages would work for all specs.

A couple of additional notes:
* The index page is not part of the list of pages.
* Only the URLs of the pages are returned.