Expose filename of the index page (#124)
This update adds a `filename` property to `release` and `nightly` to report the
actual filename of the index page, with a view to allowing external code to
identify all links to self.

The code extracts the filename from the URL if it already appears there. For
instance, it extracts the filename `identity.html` from
`https://w3c.github.io/webrtc-identity/identity.html`.

If the filename is not in the URL, the code uses a trial and error approach: it
appends `Overview.html` (most common filename) to the URL and fetches the
resulting URL. If that fails, it tries with `index.html` (second most common
filename). If that also fails, it tries with `cover.html` (which is essentially
only used for the main CSS spec). So far, this is enough to cover all existing
specs. So far, this is enough to cover all existing specs but note the code also looks at the Content-Location header as last resort.

Fixes #16.