Use previous filename when it cannot be determined (#131)
CSS drafts server does not seem to like it when it receives too many requests
in a row (see #130). This makes the `filename` info switch back and forth
between `null` and its correct value.

To avoid this, this update re-uses the previous information as a fallback. The
update also makes the code sleep 50ms between requests (this means the
generation takes about 20s more than it used to).

Also, note that the generation logic had to be changed slightly to avoid having
a `> index.json` command in `package.json`, since that was resetting the
`index.json` file right away (and in particular before the code could load it).