Use cg-report as a stronger hint of relevance
So far, find-specs would only use cg-report as a hint to look if the repo has a home page or not, and if not, would simply check if a generated page exists.

With this change, cg-report causes repos to be automatically flagged as relevant if they have a cg-report type. This identifies a bunch of new repos (and so closes #150). That being said, since these are repos which don't have github generated pages, it's not clear whether they're really readdy for inclusion in browser-specs.