Get current specification from W3C API (#186)
The current specification in a series of specifications was the last level that
was not a delta spec, unless otherwise specified in specs.json.

This update extends the definition as follows: for W3C specifications, the
current specification becomes the one returned by the W3C API, unless otherwise
specified in specs.json, and provided that the current specification returned by
the W3C API is in the list.

In turn, this allows us to drop the manual "current" entries in specs.json. They
are no longer needed, the W3C API returns the right info for them.

Fixes #176.