Add periodic check on base URL
The base URL of a specification should either match its release URL if there is
one or its nightly URL otherwise.

This update adds a `check-base-url` script and a weekly GitHub action, run on
Mondays at 00:30 (to avoid collision with the monitor action), to make sure
that this is actually the case. The action creates an issue when there are
specifications to check.