Expose path to source of nightly spec in sourcePath
This update makes the build process compute the relative path to the source of
the nightly version of each spec at the HEAD of the default branch of the spec
repository. That information gets exposed in the `nightly.sourcePath` property.

The code retrieves and parses the tree of the repository from the GitHub API,
looking for common patterns (such as `index.bs`). The code follows the same
logic as in the Web IDL auto updater:
https://github.com/saschanaz/webidl-updater/blob/main/src/find-source-path.js

As opposed to the approach taken in the auto updater, the code here does not
attempt to differentiate between types of specs (because that does not seem
needed) and does not try to deal with exceptions-to-the-rule which rather need
to be handled by specifying `nightly.sourcePath` in `specs.json`. In practice,
that is needed for 17 specifications out of 379.

A valid GitHub authentication token is needed not to end up being rate limited
by GitHub. The same authentication token is now also used to retrieve the
case-sensitive version of repository owners.

The `sourcePath` property should always be set, and tests will fail if one is
missing.

Close #173.