Exclude symlinks when looking for source path
Some WebRTC specs have an `index.html` file that is a symlink to the actual
source. Detecting symlinks through the GitHub API is doable (through `mode`).
This update excludes them.

It also adds a new rule to look for `${spec.shortname}.html` file... which is
useful to cover a couple of these WebRTC cases, but unfortunately triggers a
false positive on the Gamepad spec, which has to be hardcoded. The rule still
seems useful because that pattern could be used again in the future.

Note that, ideally, we would resolve the symlink instead of excluding it but
there is simply no way to retrieve the path to the targeted for file symlinks
from the GitHub API. That's only possible for folder symlinks:
https://developer.github.com/v3/repos/contents/#response-if-content-is-a-symlink