Avoid running tests twice on internal pull requests
There were two actions that ran tests: one that reacted to `push` and that ran
all tests, and another that reacted to `pull_request` and that only ran tests
that do not need a W3C API token, since that secret is inaccessible from PRs
created from another repository.

One net result is that tests ran twice for pull requests created from this
repository (because they are both a `push` and a `pull_request`).

This update gets back to only one action and triggers the right set of tests
depending on whether the action can access the secret. This ensures that we
run as many tests as possible against pull requests, and only run them once.