Expose test repository and paths for each spec
This update exposes a new `tests` property that details, when that is possible:
- the URL of the repository that contains the test suite for the spec
- the actual paths in that repository that contain the test suite
- if needed, the sub-paths to exclude from the list of paths because they
actually contain tests for another specification

The information is based on META.yml files in the Web Platform Tests (WPT)
repository.

A few exceptions need to be handled in specs.json:
- the ECMAScript spec
- WebGL 1 and 2
- The WebAssembly Core spec, which uses its own repository
- The WebGPU spec. The spec does have a folder under WPT but it is empty for now
as the actual test suite is developed in the gpuweb/cts repository.

A few notes:
- The code associates WebGL extension specs with the main WebGL test
repository. The association could be more fine-grained.
- Similarly, the code skips TC39 proposals, which may or may not have tests
in the main ECMAScript test repository.
- The code clones the WPT repository to a local cache folder because the repo
is too large to process it with the GitHub API (we'd be quickly rate limited,
even with an API token). To avoid creating too many files, the checkout is
restricted to META.yml files.
- `testPaths` is an array that contains one entry in most cases, except for DOM,
HTML and a couple of other specs. An alternative would be to use a space- or
comma-separated list but then that would require parsing the values.
- The code does not include provisions to detect changes. For instance, it would
seem useful to review the ~50 specs that don't have a test suite once in a while
to make sure that they still don't have one. To be done separately.