Always check repo against ignore/monitor lists
The tool that looks for potential new specs did not always check candidate
repos against the list of repos in `ignore.json` and `monitor.json`. This
update forces the code to look at these lists when it considers whether a
candidate is in scope.

Note explicit checks on the initial list of repos are no longer stricto senso
necessary and have been dropped.

Done to address #229, which should not have contained EPUB entries because the
underlying repo is in the ignore list.