Add DNT, SVG 1.1, and Manifest App info specs (#285)
This adds three requested specs that sit at the boundary of spec selection
criteria for browser-specs:

1. DNT, requested in #281
2. SVG 1.1, requested in #273
3. Web App Manifest - App info, requested in #283

The SVG 1.1 entry uses SVG 2 as nightly URL for lack of a better alternative.

JSON schema has to be slightly adjusted to account for the "-" in
"tracking-dnt.html". Also, a test started to fail because "hr-time-3" is now
the default level for HR Time.