Add "releaseUrl" and "nightlyUrl" to "series" objects
This adds a `releaseUrl` and `nightlyUrl` properties to `series` objects that
target the unversioned URLs.

The properties are set when possible. In many cases, typically when the spec is
a non-leveled spec, they are set to the same URLs as the `release.url` and
`nightly.url` properties.

Generation code remains fairly generic. This seems to work well.

Fixes #282.