Monitor Intl.NumberFormat, fix shortnames of Intl.DisplayNames
The Intl.NumberFormat proposal may be at stage 3 but it only exists as a README
for now (and README claims it is at stage 2...). Our code cannot set the
`sourcePath` as a result and it seems premature to add the spec as a result.

Also, the `tc39-` prefix was missing from the Intl.DisplayNames shortname and
the series shortname didn't look too good. Code could probably be fixed to
support `-vXX` URL suffixes, but then there is only one occurrence of such a
URL for the time being, so handling it directly in specs.json.

Close #388.