Add W3C Process and Patent Policy document, adjust tools and tests (#391)
See #390. The specs don't belong to browser-specs in theory and we may want to
separate or flag these documents somehow later on. In the meantime, they define
useful terms that other specs may want to reference.

This required a few fixes to the `fetch-groups.js` utility which crashed on
these entries (even though their groups are explicitly set in `specs.json`).
Also, the Patent Policy document is not developed in any GitHub repository for
now, so the tests that looked for `repository` on all non IETF specs needed to
be adjusted as well.