Add WICG/FedCM to monitor list, w3c/imsc-hrm to ignore list (#415)
The Federated Credential Management API is the new name of the WebID spec. Still
at the exploratory phase, although there seems to have been lots of recent
progress on the spec. No entry in Chrome status yet though.

The IMSC spec is not in browser-specs so ignoring IMSC-HRM as well.

Fix #414.