Monitor Model Element and Chinese Layout Gap Analysis
Raised in https://github.com/w3c/browser-specs/issues/453

The Model Element spec is a new proposal for WebXR developed in the Immersive
Web CG. Spec is a placeholder for now.

The Chinese Layout Gap Analysis is on the Note track and should not have been
detected as a new spec, but it has been incorrectly published on the Rec track.
Status should get updated any time soon but it's been 3 weeks already, so adding
it to monitor list for now (we should be able to remove the entry from the
monitor list next time we check it).