Update URL of Contact Picker API
The spec was moved to a W3C repo without adding a redirect, which made a cool
URI sad and broke the build. This update switches to the new URL. It also
slightly improves error reporting to log the offending URL when a 404 occurs.

Also note ongoing PR to add a redirect:
https://github.com/WICG/wicg.github.io/pull/14