Detect draft CSS specs server hiccups
The server used to serve draft CSS specs sometimes returns the contents of the
directory instead of the actual spec. The code happily considered that the
content received in such cases was an actual spec and updated the spec title
accordingly, e.g. to "Index of css-animations-2".

This update makes the code throw an error when this happens. This will make the
job fail with an error even though there is nothing that we can really do about
it, but that seems preferable to silently committing invalid titles to
`index.json`.