Add series titles to all specs
This adds a `title` and a `shortTitle` property to `series` that contain
version-less versions of the spec's title and short title. The version-less
version is retrieved from the W3C API for /TR specs, and computed by dropping
the level from the spec's title otherwise.

When a `shortTitle` is set explicitly for the spec in `specs.json`, that
title is used to compute the series' short title. This makes it possible to
compute short titles such as "WebRTC" or "ECMAScript" instead of more verbose
ones.

Close #359.