Always drop level from series shortTitle
Previous update introduced `series.shortTitle` but the level was incorrectly
kept at the end of the series short title in some cases.

Explicit short titles that get set in specs.json also need to be consistent for
the code to generate consistent series short titles.

Note that both issues are correctly detected and reported by the
"has consistent series info" test in `test/index.js`, which makes sure that the
series info is the same for all specs in a series.