Fix minor version bump logic
When a `return` statement returns something, that something must appear on the
same line... The `isMinorBumpNeeded` function always returned a result that
evaluated to `false` in practice.

The function now uses the `xxx@latest` branches as basis for comparison so as
to compare the current list with the list that was last published.