Fix release workflow (#502)
* Handle "static" files separately in diff

The `README.md` and `LICENSE.md` should not change much and cannot be easily
integrated in the diff without triggering markdown formatting issues. They are
now reported separately at the beginning of the diff as "Files have changed".

* Fix minor version bump logic

When a `return` statement returns something, that something must appear on the
same line... The `isMinorBumpNeeded` function always returned a result that
evaluated to `false` in practice.

The function now uses the `xxx@latest` branches as basis for comparison so as
to compare the current list with the list that was last published.