Minor version bump script: use full remote ref name
This update makes the script look for the fullname of the remote ref for
`web-specs@latest` and `browser-specs@latest`. On top of being more accurate
(no point basing the logic on a commit that only exists locally), this is
needed when the repo is cloned from scratch and does not have local versions
of these branches, which is typically what happens in a job.

This update also drops the escaping which I don't think is needed.

See additional context in:
https://github.com/w3c/browser-specs/pull/504#issuecomment-1034062322