Add build-diff CLI to get a glimpse of updated data
This adds a new command-line tool `src/build-diff.js` that can be used to
generate only the specs that have changed in `specs.json` between two commits.

For instance, to compute the diff between the current updated (but not
committed) list of specs in `specs.json` and the one at HEAD, run:

```
node src/build-diff.js working HEAD
```

... where `working` is a special ref that means: the version of `specs.json` in
your working folder, and `HEAD` is Git's HEAD. Any Git reference may be used,
such as a commit ID (SHA).

The tool outputs a JSON object with `added`, `updated` and `deleted` properties.
If you rather want it to spit out the new `index.json` file, add a `full`
parameter, as in:

```
node src/build-diff.js working HEAD full
```

... which you could redirect to `index.json` and follow with a `git diff` to
study the exact differences.

Important: The tool only generates the specs that have changed in `specs.json`.
Running a full build could perhaps update other parts as well (e.g. because
spec info has changed, even though `specs.json` has not).

The tool is intended to be used for debugging. It is also intended to be used in
the future as basis to compute a preview of the diff in a pull request, since a
pull request usually does not contain `index.json` (see #17).

This update also re-writes `src/build-index.js` to wrap the code in async
functions that can be parameterized with a list of specs (instead of using
hardcoded values).