Add new composition level: "fork"
This new composition level is meant for specs that are created as temporary
forks of base specifications. First (and only so far) spec in that category is
the "WebAssembly JavaScript Interface: Exception Handling" specification, as
described in https://github.com/w3c/browser-specs/issues/508.

This will require a major version bump as this changes the semantics of the
`seriesComposition` property.