Add forkOf and forks properties
Take 2 on handling temporary forks of base specifications, as described in #508.
In the final list, forks will be identified both because they will have a
`seriesComposition` property set to `"fork"`, and because they will have a
`forkOf` property that contains the shortname of the spec that they are a fork
of.

The shortname of forks will always be something like
`[base shortname]-fork-[shortname]` (no longer such a short name, isn't it?)

Specifications that have forks will now also have a `forks` property that lists
the shortnames of the forks.

Forks no longer take part in `seriesPrevious`, `seriesNext` links.