Make all jobs use Node.js v16 and "npm ci"
Jobs were still using Node.js v14, which works fine for the scripts, but comes
with a version of npm that does not support version 2 of the file format for
`package-lock.json`. On top of that, a couple of scripts were running
`npm install` instead of `npm ci`, and former updates `package-lock.json` when
needed. This could trigger the creation of pull requests that included the
`package-lock.json` file whereas they shouldn't as in #539.

This update makes all jobs switch to Node.js v16 (and ubuntu-latest, which seems
like a safe setting to use) and use `npm ci` to install dependencies.